netfilter: uapi: Merge same xt_*.c files
Merge same xt_*.h files which has the same name in upper
and lower case format.
